BugTrack/734: Apache 1.3.32 or more says an error 'Regex could not be compiled'(Apache PR: 28218) for 'i' option of FilesMatch.
Ignore-case function for Windows had been included already:
http://cvs.apache.org/viewcvs.cgi/apache-1.3/src/main/http_core.c
http://cvs.apache.org/viewcvs.cgi/httpd-2.0/server/core.c
> #ifdef WIN32
> #define USE_ICASE REG_ICASE
> #else
> #define USE_ICASE 0
> #endif

The 'i' option seems not to work for only Apache 1.3.x
(but Apache 2.0.x) bacause of difference of regex
imprementation (Info from Yocchi).

So this option here, is redundant, and become a trap
causing Internal Server Error with Apache 1.3.x for Unix.