BugTrack2/62: Do remove the whole design, 'Showing TrackBack-ping list by html'. This part suggets us here must be a wrong way to 'implement/test' 'new'  feature for our/your product. You guys must not think copying other's original work even a bit without suitable permisson or license.
(1) Here is the same CSS file of Movable Type related site. That had been copied from TrackBack development site. But the license of the file seems not clear for the use with OpenSource product.
(2) So, here are the same CSS classes within (1), the same artistic design(look), and the same information structure of (1).
(3) And, here are the same and/or translated strings of 'Movable Type'(A product of Six Apart) used many at the Internet. This is the worst one.

Strongly note that 'implementing/testing something new' is not 'copying design and/or copying strings and/or replacing strings with translation'.